refactor: samples
