+++ samples/books.ts
@@ -1,0 +1,5 @@
+import book from "./book";
+
+const books = [...Array(10)].map((_value, id) => ({ ...book, id }));
+
+export default books;
