+++ samples/index.ts
@@ -1,0 +1,6 @@
+export { default as resouce } from "./resource";
+export { default as topic } from "./topic";
+export { default as section } from "./section";
+export { default as sections } from "./sections";
+export { default as book } from "./book";
+export { default as books } from "./books";
