+++ samples/resource.ts
@@ -1,0 +1,8 @@
+const resource = {
+  id: 1,
+  videoId: 1,
+  url: "YzVwrvbz_XA",
+  details: {},
+};
+
+export default resource;
