+++ samples/section.ts
@@ -1,0 +1,9 @@
+import topic from "./topic";
+
+const section = {
+  id: 1,
+  name: "情報のデジタルコンテンツ化",
+  topics: [topic],
+};
+
+export default section;
