+++ samples/topic.ts
@@ -1,11 +1,5 @@
 import outdent from "outdent";
-
-const resource = {
-  id: 1,
-  videoId: 1,
-  url: "YzVwrvbz_XA",
-  details: {},
-};
+import resource from "./resource";
 
 const topic = {
   id: 1,
@@ -35,33 +29,4 @@ const topic = {
   resource,
 };
 
-const sections = [
-  {
-    id: 1,
-    name: "情報のデジタルコンテンツ化",
-    topics: [topic],
-  },
-  {
-    id: 2,
-    name: null,
-    topics: [topic],
-  },
-  {
-    id: 3,
-    name: "デジタルとアナログの相違点",
-    topics: [],
-  },
-];
-
-const book = {
-  id: 1,
-  name: "コンピュータ・サイエンス",
-  author: { name: "山田太郎" },
-  createdAt: new Date(),
-  updatedAt: new Date(),
-  sections,
-};
-
-const props = { book };
-
-export default props;
+export default topic;
